+++ frontend/src/menu/Main.tsx
@@ -1,6 +1,6 @@
-import { useApolloClient, useMutation } from "@apollo/client"
+import { useApolloClient, useMutation, useQuery } from "@apollo/client"
 import { Divider, List } from "@material-ui/core"
-import { Brightness2 as Brightness2Icon, BrightnessHigh as BrightnessHighIcon, ChevronRight as ChevronRightIcon, Event as EventIcon, ExitToApp as ExitToAppIcon, GitHub as GitHubIcon, LockOpen as LockOpenIcon, Notifications as NotificationsIcon, PersonAdd as PersonAddIcon, VideogameAsset as VideogameAssetIcon } from "@material-ui/icons"
+import { Brightness2 as Brightness2Icon, BrightnessHigh as BrightnessHighIcon, ChevronRight as ChevronRightIcon, Event as EventIcon, ExitToApp as ExitToAppIcon, GitHub as GitHubIcon, LockOpen as LockOpenIcon, MeetingRoom as MeetingRoomIcon, Notifications as NotificationsIcon, PersonAdd as PersonAddIcon, VideogameAsset as VideogameAssetIcon } from "@material-ui/icons"
 import gql from "graphql-tag"
 import { useState, CSSProperties, useEffect } from "react"
 import { useHistory } from "react-router"
@@ -24,10 +24,12 @@ export default function Main(props: Props) {
 	const { isLoggedIn } = useWhoami()
 	const client = useApolloClient()
 	const [logout] = useMutation(LOGOUT)
+	const [leaveLobby] = useMutation(LEAVE_LOBBY)
 	const { ref } = useResizeObserver({
 		onResize: ({ height }) => props.onResize(height)
 	})
 	const [version, setVersion] = useState<string>()
+	const { data: lobbyResult } = useQuery<LobbyResult>(LOBBY, { fetchPolicy: "cache-first" })
 
 	// get version number
 	useEffect(() => {
@@ -58,19 +60,34 @@ export default function Main(props: Props) {
 						<>
 							<Item
 								onClick={() => {
-									history.push("/lobby")
-									props.onClose()
-								}}
-								icon={<VideogameAssetIcon />}
-								title="Spiel" />
-
-							<Item
-								onClick={() => {
 									history.push("/calendar")
 									props.onClose()
 								}}
 								icon={<EventIcon />}
 								title="Kalendar" />
+
+							{
+								lobbyResult?.lobby == null ?
+									<Item
+										onClick={() => {
+											history.push("/lobby")
+											props.onClose()
+										}}
+										icon={<VideogameAssetIcon />}
+										title="Spiel" /> :
+									<Item
+										onClick={async () => {
+											try {
+												await leaveLobby()
+											} catch (err) {
+												console.error(err)
+											}
+											history.push("/lobby")
+											props.onClose()
+										}}
+										icon={<MeetingRoomIcon />}
+										title="Verlassen" />
+							}
 						</> :
 						<>
 							<Item
@@ -135,5 +152,23 @@ export default function Main(props: Props) {
 const LOGOUT = gql`
 	mutation {
 		logout
+	}
+`
+
+const LEAVE_LOBBY = gql`
+	mutation {
+		leaveLobby
+	}
+`
+
+interface LobbyResult {
+	lobby: {} | null
+}
+
+const LOBBY = gql`
+	query {
+		lobby {
+			__typename
+		}
 	}
 `